Move KeyPair to crypto
