+++ crate/crypto/src/lib.rs
@@ -1,3 +1,5 @@
+#![feature(slice_take)]
+
 pub mod cover_crypt;
 pub mod dh_shared_keys;
 pub mod elliptic_curves;
@@ -9,7 +11,10 @@ pub mod rsa;
 pub mod symmetric;
 pub mod wrap;
 
-use cosmian_kmip::kmip::kmip_operations::{Decrypt, DecryptResponse, Encrypt, EncryptResponse};
+use cosmian_kmip::kmip::{
+    kmip_objects::Object,
+    kmip_operations::{Decrypt, DecryptResponse, Encrypt, EncryptResponse},
+};
 pub use elliptic_curves::Q_LENGTH_BITS;
 use error::KmsCryptoError;
 
@@ -25,4 +30,39 @@ impl<T: EncryptionSystem + ?Sized> EncryptionSystem for Box<T> {
 
 pub trait DecryptionSystem {
     fn decrypt(&self, request: &Decrypt) -> Result<DecryptResponse, KmsCryptoError>;
+}
+
+/// A `KeyPair` is a tuple `(Object::PrivateKey, Object::PublicKey)`
+///
+/// Note: this object does not exist in the KMIP specs,
+/// hence its definition here
+pub struct KeyPair(pub (Object, Object));
+impl KeyPair {
+    /// Create a new `KeyPair` from a private and public key
+    #[must_use]
+    pub fn new(private_key: Object, public_key: Object) -> Self {
+        Self((private_key, public_key))
+    }
+
+    /// Get the private key
+    #[must_use]
+    pub fn private_key(&self) -> &Object {
+        &self.0.0
+    }
+
+    /// Get the public key
+    #[must_use]
+    pub fn public_key(&self) -> &Object {
+        &self.0.1
+    }
+
+    /// Get the private key
+    pub fn private_key_mut(&mut self) -> &mut Object {
+        &mut self.0.0
+    }
+
+    /// Get the public key
+    pub fn public_key_mut(&mut self) -> &mut Object {
+        &mut self.0.1
+    }
 }
