+++ crate/server/src/core/cover_crypt/create_user_decryption_key.rs
@@ -4,14 +4,14 @@ use cosmian_kmip::kmip::{
     kmip_operations::{Create, CreateKeyPair, ErrorReason, Get},
     kmip_types::{Attributes, KeyFormatType, StateEnumeration},
 };
-use cosmian_kms_crypto::cover_crypt::{
-    attributes::{access_policy_from_attributes, policy_from_attributes},
-    user_key::UserDecryptionKeysHandler,
-};
-use cosmian_kms_utils::{
-    access::{ExtraDatabaseParams, ObjectOperationType},
+use cosmian_kms_crypto::{
+    cover_crypt::{
+        attributes::{access_policy_from_attributes, policy_from_attributes},
+        user_key::UserDecryptionKeysHandler,
+    },
     KeyPair,
 };
+use cosmian_kms_utils::access::{ExtraDatabaseParams, ObjectOperationType};
 
 use super::KMS;
 use crate::{
